+++ test/mq_client_api.test.js
@@ -3,14 +3,17 @@
 const assert = require('assert');
 const address = require('address');
 const httpclient = require('urllib');
+const config = require('../example/config');
 const MQClientAPI = require('../lib/mq_client_api');
 
-const TOPIC = 'GXCSOCCER';
+const TOPIC = config.topic;
 
 describe('test/mq_client_api.test.js', function() {
   let client;
   before(() => {
-    client = new MQClientAPI(Object.assign({ httpclient }, require('../example/config')));
+    client = new MQClientAPI(Object.assign({ httpclient }, Object.assign({}, config, {
+      nameSrv: [ 'onsaddr.mq-internet-access.mq-internet.aliyuncs.com:80' ],
+    })));
     return client.ready();
   });